A simple framework for a lexer
This doesn't really "lex" anything except for end of input right now, but it shows how the overall structure will work. Basically, the parser will ask the lexer for the next token when it needs it. The lexer is designed to be super-simple; basically, it will only lex the overall "structure" of the language. As a consequence of this, almost everything it lexes that isn't a keyword or structural language symbol will be lexed as an "ident". It will be up to the parser, semantic analyzer, and code generator to further qualify and validate identifiers.

Another lexer feature is that it retains tokens for whitespace and comments. This is because we will emit the corresponding C code with comments and formatting that came from the original source file.

To get the file contents, we simply read the whole file into memory before lexing. We aren't going to bother optimizing this kind of thing right now.