Report token info; lex more symbols and start lexing identifiers
We now lex the equals sign and whitespace. We also start to lex identifiers. We wrote an fprint_token() routine that output the token information in JSON-style formatting. This gives us something we can input to processing tools for analysis and also lets us spot bugs in our lexer, which we can immediately see now that we are explicitly printing lexed content. Bugs we currently see:

- We're somewhat inconsistent with where location.range_end is. Is it one AFTER the lexeme or at the very end?
- We can only lexing identifiers in chunks of two characters at a time
- If we go into two-char lexing mode, we treat the second character incorrectly if it's the start of a valid token