Add a simple semantic analyzer
In this commit we build out the framework for a semantic analyzer. The first iteration just checks for identifier mentions in expressions and makes sure they are defined in one of the enclosing lexical scopes.

- Add ASTLIST_FOREACH macro
- Add asserts
- Create the notion of a "spelling" (a start and end location in a buffer) and make operations for them
- Add spelling as an alternate way of getting at the source location range with a union
- Add parent relationships to the AST
- Simple notion of "scope" as being a node that can contain definitions
- Demonstrates using a visitor to gather data then analyzing that data
- Fixes bugs in parser and in ast_list_add
- Extracts printing last node into a helper function for easier use outside of the print visitor

Does NOT yet demonstrate tagging the AST with meaningful information during the analysis phase. However, if you attempt to use an undeclared identifier it will correctly error out. This is necessary before moving to code generation because it should be a rule that the generated code always compiles without errors or warnings (if it doesn't, then we missed something in the analyzer or are generating incorrect code).