Add visit_data temp pointer to ASTBase (prep work for further functionality)
More complex passes will need to store their own data in each AST node. This functionality provides a uniform way to do that. If a visitor wants to attach data to a node, it does so with a malloc call. After the full pass is complete (usually after the ast_visit() call), the pass must call ast_visit_data_clean() to free up the memory and set the pointers back to NULL. This is essential to prevent subsequent passes from seeing a non-NULL visit_data ptr and assuming the type to be its own.